refactor(wallet): rename ActiveStakePublic files to PublicStake
